Fix for no household input files.
