+++ test-data/envs/small-no-household/output/link_ids.csv
@@ -1,0 +1,28 @@
+LINK_ID,hospital_001,hospital_002,hospital_003
+b2dbcf83-2935-11ec-b9bc-0433c29f7e36,0,,
+b2dbcf84-2935-11ec-a625-0433c29f7e36,1,,
+b2dbcf85-2935-11ec-a579-0433c29f7e36,2,,
+b2dbcf86-2935-11ec-959c-0433c29f7e36,3,,
+b2dbcf87-2935-11ec-83da-0433c29f7e36,4,,
+b2dbcf88-2935-11ec-9d5c-0433c29f7e36,5,,
+b2dbcf89-2935-11ec-ad7e-0433c29f7e36,6,,
+b2dbcf8a-2935-11ec-8931-0433c29f7e36,7,,
+b2dbcf8b-2935-11ec-b23a-0433c29f7e36,8,,
+b2dbcf8c-2935-11ec-8c73-0433c29f7e36,,0,
+b2dbcf8d-2935-11ec-ac3f-0433c29f7e36,,1,
+b2dbcf8e-2935-11ec-806f-0433c29f7e36,,2,
+b2dbcf8f-2935-11ec-8495-0433c29f7e36,,3,
+b2dbcf90-2935-11ec-a89e-0433c29f7e36,,4,
+b2dbcf91-2935-11ec-89df-0433c29f7e36,,5,
+b2dbcf92-2935-11ec-87d3-0433c29f7e36,,6,
+b2dbcf93-2935-11ec-bffc-0433c29f7e36,,7,
+b2dbcf94-2935-11ec-91cd-0433c29f7e36,,8,
+b2dbcf95-2935-11ec-a1cb-0433c29f7e36,,,0
+b2dbcf96-2935-11ec-abc4-0433c29f7e36,,,1
+b2dbcf97-2935-11ec-a39f-0433c29f7e36,,,2
+b2dbcf98-2935-11ec-8694-0433c29f7e36,,,3
+b2dbf6ab-2935-11ec-90f8-0433c29f7e36,,,4
+b2dbf6ac-2935-11ec-80fe-0433c29f7e36,,,5
+b2dbf6ad-2935-11ec-876e-0433c29f7e36,,,6
+b2dbf6ae-2935-11ec-bd6f-0433c29f7e36,,,7
+b2dbf6af-2935-11ec-b27b-0433c29f7e36,,,8
