+++ test-data/envs/small-no-household/schema/blocking-schema/lambda.json
@@ -1,0 +1,13 @@
+{
+    "type": "lambda-fold",
+    "version": 1,
+    "config": {
+        "blocking-features": [6],
+        "Lambda": 3,
+        "bf-len": 64,
+        "num-hash-funcs": 3,
+        "K": 5,
+        "input-clks": true,
+        "random_state": 0
+    }
+}
