+++ test-data/envs/small-no-household/schema/blocking-schema/p-sig.json
@@ -1,0 +1,27 @@
+{
+ "type": "p-sig",
+ "version": 1,
+ "config": {
+     "blocking_features": [1],
+     "filter": {
+         "type": "ratio",
+         "max": 0.02,
+         "min": 0.00
+     },
+     "blocking-filter": {
+         "type": "bloom filter",
+         "number-hash-functions": 4,
+         "bf-len": 2048
+     },
+     "signatureSpecs": [
+         [
+              {"type": "characters-at", "config": {"pos": [0]}, "feature-idx": 1},
+              {"type": "characters-at", "config": {"pos": [0]}, "feature-idx": 2}
+         ],
+         [
+             {"type": "metaphone", "feature-idx": 1},
+             {"type": "metaphone", "feature-idx": 2}
+         ]
+     ]
+   }
+}
