--- test-data/envs/small/output/link_ids.csv
@@ -1,28 +1,0 @@
-LINK_ID,hospital_001,hospital_002,hospital_003
-1c3dce55-292c-11ec-a2f2-0433c29f7e36,0,,
-1c3dce56-292c-11ec-95d6-0433c29f7e36,1,,
-1c3dce57-292c-11ec-87a1-0433c29f7e36,2,,
-1c3dce58-292c-11ec-b9af-0433c29f7e36,3,,
-1c3dce59-292c-11ec-b5e3-0433c29f7e36,4,,
-1c3dce5a-292c-11ec-b223-0433c29f7e36,5,,
-1c3dce5b-292c-11ec-b5e7-0433c29f7e36,6,,
-1c3dce5c-292c-11ec-b5dd-0433c29f7e36,7,,
-1c3dce5d-292c-11ec-a182-0433c29f7e36,8,,
-1c3dce5e-292c-11ec-a8c1-0433c29f7e36,,0,
-1c3dce5f-292c-11ec-9243-0433c29f7e36,,1,
-1c3dce60-292c-11ec-a60e-0433c29f7e36,,2,
-1c3dce61-292c-11ec-b76d-0433c29f7e36,,3,
-1c3dce62-292c-11ec-9d1f-0433c29f7e36,,4,
-1c3dce63-292c-11ec-a7bc-0433c29f7e36,,5,
-1c3dce64-292c-11ec-b60c-0433c29f7e36,,6,
-1c3dce65-292c-11ec-b730-0433c29f7e36,,7,
-1c3dce66-292c-11ec-8230-0433c29f7e36,,8,
-1c3dce67-292c-11ec-a623-0433c29f7e36,,,0
-1c3dce68-292c-11ec-9710-0433c29f7e36,,,1
-1c3dce69-292c-11ec-90cd-0433c29f7e36,,,2
-1c3dce6a-292c-11ec-8ad5-0433c29f7e36,,,3
-1c3dce6b-292c-11ec-adc3-0433c29f7e36,,,4
-1c3dce6c-292c-11ec-8a09-0433c29f7e36,,,5
-1c3dce6d-292c-11ec-9cf7-0433c29f7e36,,,6
-1c3dce6e-292c-11ec-bcbd-0433c29f7e36,,,7
-1c3dce6f-292c-11ec-b50f-0433c29f7e36,,,8
