+++ test/linkage-agent/individual-steps/a_test_drop.py
@@ -2,14 +2,9 @@ import util.config.config_util as cu
 import util.linkage_agent.linkage_agent_util as lau
 
 
-def test_match():
+def test_drop():
     print("\nStarting test...")
     config = cu.get_config("test-data/defaults/config.json")
-    print("Doing validation...")
-#    lau.validate(config)
-    print("Doing match...")
-#    lau.match(config)
-    print("Doing link_id...")
-    lau.link_id(config)
+    lau.drop(config)
     print("Done.")
