+++ test/linkage-agent/individual-steps/d_test_link_ids.py
@@ -1,0 +1,17 @@
+import util.config.config_util as cu
+import util.linkage_agent.linkage_agent_util as lau
+
+
+def test_match():
+    print("\nStarting test...")
+    config = cu.get_config("test-data/defaults/config.json")
+    print("Doing validation...")
+    lau.validate(config)
+    print("Doing match...")
+    lau.match(config)
+    print("Doing link_id...")
+    lau.link_id(config)
+    print("Doing drop...")
+    lau.drop(config)
+    print("Done.")
+
