+++ test/linkage-agent/no-household/config-no-households.json
@@ -1,0 +1,24 @@
+{
+  "systems": [
+    "hospital_001",
+    "hospital_002",
+    "hospital_003"
+  ],
+  "projects": [
+    "name-sex-dob-phone",
+    "name-sex-dob-zip",
+    "name-sex-dob-parents",
+    "name-sex-dob-addr"
+  ],
+  "schema_folder": "test-data/envs/small-no-household/schema",
+  "inbox_folder": "test-data/envs/small-no-household/inbox",
+  "matching_results_folder": "test-data/envs/small-no-household/output",
+  "output_folder": "test-data/envs/small-no-household/output",
+  "entity_service_url": "http://localhost:8851/api/v1",
+  "matching_threshold": 0.75,
+  "mongo_uri": "localhost:27017",
+  "blocked": false,
+  "blocking_schema": "test-data/envs/small-no-household/schema/name-sex-dob-addr.json",
+  "household_match": false,
+  "household_schema": "test-data/envs/small-no-household/schema/name-sex-dob-parents.json"
+}
